Remove # when listing webhook ids
As a user I didn't know if the # was part of the webhook ID or not. (Actually I first thought that the 1.) on the left side was the hook id)

When I double-click on the ID to select it in the terminal to copy it and paste it after `particle webhook delete` I get an error because the # was selected as part of the same word. I have to carefully select all characters except the first # when copying the webhook id.

It would be easier to delete hooks if I could simply double-click on the id to copy it.